Fix bug in bug fix for bug fix

+++ models/IGM/model.py
@@ -55,7 +55,7 @@ class Net(nn.Module):
             data_dict['loss'] = F.kl_div(y_src, y_tgt_pm) - F.kl_div(y_src, y_tgt_rand)
         lab_src = y_src.argmax(1)
         lab_tgt = y_tgt.argmax(1)
-        perm_vec = (lab_tgt.unsqueeze(-2) == lab_src.unsqueeze(-1)).argmax(-1)
+        perm_vec = (lab_tgt.unsqueeze(-2) == lab_src.unsqueeze(-1)).float().argmax(-1)
         perm_mat = torch.eye(lab_tgt.shape[-1]).to(y_src)[perm_vec]
         data_dict['perm_mat'] = perm_mat
         data_dict['ds_mat'] = perm_mat
